-integrate everything fetch news api in news fragment -add bottom navigation functionality -update fragments contents -add icons
