+++ .idea/libraries/Gradle__androidx_activity_activity_1_3_1_aar.xml
@@ -1,0 +1,16 @@
+<component name="libraryTable">
+  <library name="Gradle: androidx.activity:activity:1.3.1@aar">
+    <ANNOTATIONS>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/25bfc18c2b39130cc93ee66137052dc8/transformed/activity-1.3.1/annotations.zip!/" />
+    </ANNOTATIONS>
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/25bfc18c2b39130cc93ee66137052dc8/transformed/activity-1.3.1/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/25bfc18c2b39130cc93ee66137052dc8/transformed/activity-1.3.1/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/25bfc18c2b39130cc93ee66137052dc8/transformed/activity-1.3.1/AndroidManifest.xml" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.activity/activity/1.3.1/47b35d483b15e1738de81d26b51ccfc6e2e7b5f8/activity-1.3.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
