+++ .idea/libraries/Gradle__androidx_databinding_databinding_adapters_7_2_1_aar.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Gradle: androidx.databinding:databinding-adapters:7.2.1@aar">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/d056f9f94963c93e4b2b0fc3b4d7dab1/transformed/databinding-adapters-7.2.1/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/d056f9f94963c93e4b2b0fc3b4d7dab1/transformed/databinding-adapters-7.2.1/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/d056f9f94963c93e4b2b0fc3b4d7dab1/transformed/databinding-adapters-7.2.1/AndroidManifest.xml" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.databinding/databinding-adapters/7.2.1/6cc7bb80e41d8e0a4809878f2c2194b29d924123/databinding-adapters-7.2.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
