+++ .idea/libraries/Gradle__androidx_databinding_databinding_common_7_2_1.xml
@@ -1,0 +1,11 @@
+<component name="libraryTable">
+  <library name="Gradle: androidx.databinding:databinding-common:7.2.1">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.databinding/databinding-common/7.2.1/cb53bd625b7a99cf218f46634d07d1886441288/databinding-common-7.2.1.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.databinding/databinding-common/7.2.1/11a1cbee1a967c4db030fd5fbf71656ca4dd86c3/databinding-common-7.2.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
