+++ .idea/libraries/Gradle__androidx_databinding_databinding_ktx_7_2_1_aar.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Gradle: androidx.databinding:databinding-ktx:7.2.1@aar">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/0958d7f49fcf238e3db46c7ba55937d0/transformed/databinding-ktx-7.2.1/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/0958d7f49fcf238e3db46c7ba55937d0/transformed/databinding-ktx-7.2.1/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/0958d7f49fcf238e3db46c7ba55937d0/transformed/databinding-ktx-7.2.1/AndroidManifest.xml" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.databinding/databinding-ktx/7.2.1/2b4648dc63d765da73c0ece136c1c49f779a65e/databinding-ktx-7.2.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
