+++ .idea/libraries/Gradle__androidx_databinding_databinding_runtime_7_2_1_aar.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Gradle: androidx.databinding:databinding-runtime:7.2.1@aar">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/e361c5ca2d87eff0f75c2cdc9fbae78a/transformed/databinding-runtime-7.2.1/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/e361c5ca2d87eff0f75c2cdc9fbae78a/transformed/databinding-runtime-7.2.1/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/e361c5ca2d87eff0f75c2cdc9fbae78a/transformed/databinding-runtime-7.2.1/AndroidManifest.xml" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.databinding/databinding-runtime/7.2.1/41c90e7394b7e2d5180719dc821d3cb37cbf3557/databinding-runtime-7.2.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
