+++ .idea/libraries/Gradle__com_google_dagger_dagger_lint_aar_2_40_1_aar.xml
@@ -1,0 +1,11 @@
+<component name="libraryTable">
+  <library name="Gradle: com.google.dagger:dagger-lint-aar:2.40.1@aar">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/8ec357a35127d9365e237987a3d12122/transformed/dagger-lint-aar-2.40.1/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/8ec357a35127d9365e237987a3d12122/transformed/dagger-lint-aar-2.40.1/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/8ec357a35127d9365e237987a3d12122/transformed/dagger-lint-aar-2.40.1/AndroidManifest.xml" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
