+++ .idea/libraries/Gradle__com_google_dagger_hilt_android_2_40_1_aar.xml
@@ -1,0 +1,15 @@
+<component name="libraryTable">
+  <library name="Gradle: com.google.dagger:hilt-android:2.40.1@aar">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/bd474adb49c99dc6625f233811a80da2/transformed/hilt-android-2.40.1/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/bd474adb49c99dc6625f233811a80da2/transformed/hilt-android-2.40.1/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/bd474adb49c99dc6625f233811a80da2/transformed/hilt-android-2.40.1/AndroidManifest.xml" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.google.dagger/hilt-android/2.40.1/ca22d5f2383f54c6dc946cdece9c59d097a40b2a/hilt-android-2.40.1-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.google.dagger/hilt-android/2.40.1/f6347312b219cfbfdd51ccfa066a8c9c019ac7f5/hilt-android-2.40.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
