+++ .idea/misc.xml
@@ -3,9 +3,17 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="app/src/main/res/drawable/ic_comment.xml" value="0.124" />
+        <entry key="app/src/main/res/drawable/ic_home.xml" value="0.124" />
+        <entry key="app/src/main/res/drawable/ic_search.xml" value="0.124" />
+        <entry key="app/src/main/res/drawable/ic_settings.xml" value="0.124" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.175" />
         <entry key="app/src/main/res/layout/fragment_main.xml" value="0.175" />
+        <entry key="app/src/main/res/layout/fragment_news.xml" value="0.19322916666666667" />
+        <entry key="app/src/main/res/layout/fragment_search.xml" value="0.19322916666666667" />
+        <entry key="app/src/main/res/layout/fragment_settings.xml" value="0.19322916666666667" />
         <entry key="app/src/main/res/layout/item_news.xml" value="0.165" />
+        <entry key="app/src/main/res/menu/bottom_menu.xml" value="0.19322916666666667" />
       </map>
     </option>
   </component>
