+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.NewsApi">
         <activity
-            android:name=".view.MainActivity"
+            android:name=".view.activity.MainActivity"
             android:exported="true"
             tools:ignore="Instantiatable">
             <intent-filter>
