+++ app/src/main/res/menu/bottom_menu.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:title="Home" android:id="@+id/newsFragment"
+        android:icon="@drawable/ic_home"/>
+    <item android:title="Search" android:id="@+id/searchFragment"
+        android:icon="@drawable/ic_search"/>
+    <item android:title="Others" android:id="@+id/settingsFragment"
+        android:icon="@drawable/ic_settings"/>
+</menu>
